promote module to directory
